+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ### 🐞 Bug Fixes
 - Fixed an issue where NBT info could disappear from lanterns.
 - Resolved debug screen glitches and unexpected behavior.
+- Fixed an issue where lantern does not drop upon death